+++ lya_2pt/export.py
@@ -1,17 +1,29 @@
+import os.path
 from multiprocessing import Pool
 
 import fitsio
+import h5py
 import numpy as np
+import scipy.interpolate
 
+from lya_2pt.constants import ACCEPTED_BLIND_CORRELATION_TYPES, ACCEPTED_BLINDING_STRATEGIES
 from lya_2pt.output import get_coordinate_columns, get_coordinate_system, get_grid_header
 from lya_2pt.utils import parse_config
 
-accepted_options = ["export-correlation", "export-distortion", "smooth-covariance"]
+UNBLINDABLE_STRATEGIES = ["none", "desi_m2", "desi_y1", "desi_y3"]
+
+accepted_options = [
+    "export-correlation",
+    "export-distortion",
+    "smooth-covariance",
+    "blind-corr-type",
+]
 
 defaults = {
     "export-correlation": False,
     "export-distortion": False,
     "smooth-covariance": True,
+    "blind-corr-type": "",
 }
 
 
@@ -33,6 +45,12 @@ class Export:
 
         self.export_correlation = self.config.getboolean("export-correlation")
         self.export_distortion = self.config.getboolean("export-distortion")
+        self.blind_corr_type = self.config.get("blind-corr-type")
+        if self.blind_corr_type and self.blind_corr_type not in ACCEPTED_BLIND_CORRELATION_TYPES:
+            raise ValueError(
+                "Expected blind-corr-type to be one of "
+                f"{ACCEPTED_BLIND_CORRELATION_TYPES}. Found {self.blind_corr_type}."
+            )
 
     @staticmethod
     def _coordinate_system_from_header(header):
@@ -65,6 +83,120 @@ class Export:
             self.coordinate2_max = header["R_TRANS_MAX"]
             self.num_bins_coordinate1 = header["NUM_BINS_R_PAR"]
             self.num_bins_coordinate2 = header["NUM_BINS_R_TRANS"]
+
+    def _set_blinding(self, blindings):
+        """Validate and store the common blinding strategy from HEALPix files."""
+        unique_blindings = set(blindings)
+        if len(unique_blindings) != 1:
+            raise ValueError("Cannot export HEALPix files with different blinding strategies")
+
+        blinding = unique_blindings.pop()
+        if isinstance(blinding, np.number):
+            if not blinding.is_integer() or not 0 <= blinding < len(ACCEPTED_BLINDING_STRATEGIES):
+                raise ValueError(
+                    "Expected blinding strategy to be one of "
+                    f"{ACCEPTED_BLINDING_STRATEGIES}. Found {blinding}."
+                )
+            blinding = ACCEPTED_BLINDING_STRATEGIES[int(blinding)]
+        if blinding not in ACCEPTED_BLINDING_STRATEGIES:
+            raise ValueError(
+                "Expected blinding strategy to be one of "
+                f"{ACCEPTED_BLINDING_STRATEGIES}. Found {blinding}."
+            )
+        if hasattr(self, "blinding") and self.blinding != blinding:
+            raise ValueError(
+                "Cannot export correlation and distortion files with different blinding"
+            )
+        if self.coordinate_system == "R_MU" and blinding not in UNBLINDABLE_STRATEGIES:
+            raise ValueError("Blinded export is only supported for rp-rt coordinate grids")
+
+        self.blinding = blinding
+
+    def _output_is_blinded(self):
+        return self.blinding not in UNBLINDABLE_STRATEGIES
+
+    def _apply_blinding(self, xi):
+        """Apply the DESI blinding template to an aggregated correlation."""
+        blinding = self.blinding
+        if blinding in UNBLINDABLE_STRATEGIES:
+            print(f"'{blinding}' correlations are not blinded.")
+            return xi
+
+        blinding_dir = "/global/cfs/projectdirs/desi/science/lya/lya_blinding/bao/"
+        blinding_templates = {
+            "desi_dr3": {
+                "standard": "dr3_blinding_v4_standard_28_05_2026.h5",
+                "grid": "dr3_blinding_v4_regular_grid_28_05_2026.h5",
+            }
+        }
+
+        if blinding in blinding_templates:
+            print(f"Blinding using seed for {blinding}")
+        else:
+            raise ValueError(
+                f"Expected blinding to be one of {blinding_templates.keys()}. Found {blinding}."
+            )
+
+        if not self.blind_corr_type:
+            raise ValueError("Blinding requires export option blind-corr-type.")
+
+        blind_corr_type = self.blind_corr_type
+        # Match the name expected in the blinding template file.
+        if blind_corr_type == "qsoxlya":
+            blind_corr_type = "lyaxqso"
+        if blind_corr_type == "qsoxlyb":
+            blind_corr_type = "lybxqso"
+
+        # Check type of correlation and get size and regular binning.
+        if blind_corr_type in ["lyaxlya", "lyaxlyb"]:
+            corr_size = 2500
+            rp_interp_grid = np.arange(2.0, 202.0, 4)
+            rt_interp_grid = np.arange(2.0, 202.0, 4)
+        elif blind_corr_type in ["lyaxqso", "lybxqso"]:
+            corr_size = 5000
+            rp_interp_grid = np.arange(-197.99, 202.01, 4)
+            rt_interp_grid = np.arange(2.0, 202, 4)
+        else:
+            raise ValueError("Unknown correlation type: {}".format(blind_corr_type))
+
+        if corr_size == self.num_bins_coordinate1 * self.num_bins_coordinate2:
+            # Read the blinding file and get the right template.
+            blinding_filename = blinding_dir + blinding_templates[blinding]["standard"]
+        else:
+            # Read the regular-grid blinding file and get the right template.
+            blinding_filename = blinding_dir + blinding_templates[blinding]["grid"]
+
+        if not os.path.isfile(blinding_filename):
+            raise RuntimeError(
+                "Missing blinding file. Make sure you are running at"
+                " NERSC or contact picca developers"
+            )
+        with h5py.File(blinding_filename, "r") as blinding_file:
+            hex_diff = np.array(blinding_file["blinding"][blind_corr_type]).astype(str)
+        diff_grid = np.array([float.fromhex(x) for x in hex_diff])
+
+        if corr_size == self.num_bins_coordinate1 * self.num_bins_coordinate2:
+            diff = diff_grid
+        else:
+            # Interpolate the blinding template on the regular grid.
+            interp = scipy.interpolate.RectBivariateSpline(
+                rp_interp_grid,
+                rt_interp_grid,
+                diff_grid.reshape(len(rp_interp_grid), len(rt_interp_grid)),
+                kx=3,
+                ky=3,
+            )
+            diff = interp.ev(self.coordinate1, self.coordinate2)
+
+        # Check that the shapes match.
+        if np.shape(xi) != np.shape(diff):
+            raise RuntimeError(
+                "Unknown binning or wrong correlation type. Cannot blind."
+                " Please raise an issue or contact picca developers."
+            )
+
+        # Add blinding.
+        return xi + diff
 
     def run(self, global_config, settings):
         self.expected_coordinate_system = get_coordinate_system(settings)
@@ -114,6 +246,7 @@ class Export:
             results = [self._read_correlation(file) for file in files]
 
         results = np.array(results)
+        self._set_blinding(results[:, 6, :].ravel())
         self.correlations = results[:, 0, :]
         self.weights = results[:, 1, :]
         self.mean_correlation = np.sum(self.correlations * self.weights, axis=0)
@@ -131,7 +264,8 @@ class Export:
 
     def _read_correlation(self, file):
         with fitsio.FITS(file) as hdul:
-            coordinate_system = self._coordinate_system_from_header(hdul[1].read_header())
+            header = hdul[1].read_header()
+            coordinate_system = self._coordinate_system_from_header(header)
             if coordinate_system != self.coordinate_system:
                 raise ValueError("Cannot export mixed coordinate-system healpix files")
             coordinate_names, _, _ = get_coordinate_columns(coordinate_system)
@@ -140,11 +274,21 @@ class Export:
             z = hdul[1]["Z"][:]
             num_pairs = hdul[1]["NUM_PAIRS"][:]
 
-            # TODO implement blinding support
-            correlation = hdul[2]["CORRELATION"][:]
+            blinding = header["BLINDING"] if "BLINDING" in header else "none"
+            if blinding not in ACCEPTED_BLINDING_STRATEGIES:
+                raise ValueError(
+                    "Expected blinding strategy to be one of "
+                    f"{ACCEPTED_BLINDING_STRATEGIES}. Found {blinding}."
+                )
+            correlation_name = "CORRELATION"
+            if "CORRELATION_BLIND" in hdul[2].get_colnames():
+                correlation_name += "_BLIND"
+            correlation = hdul[2][correlation_name][:]
             weights = hdul[2]["WEIGHT_SUM"][:]
 
-        return correlation, weights, coordinate1, coordinate2, z, num_pairs
+        blinding_code = ACCEPTED_BLINDING_STRATEGIES.index(blinding)
+        blinding_codes = np.full(correlation.shape, blinding_code)
+        return correlation, weights, coordinate1, coordinate2, z, num_pairs, blinding_codes
 
     def read_distortion(self):
         files = np.array(list(self.healpix_dir.glob("distortion*fits*")))
@@ -160,6 +304,7 @@ class Export:
             results = [self._read_distortion(file) for file in files]
 
         results = list(results)
+        self._set_blinding([result[8] for result in results])
         self.distortion = np.array([item[0] for item in results]).sum(axis=0)
         self.dist_weights = np.array([item[1] for item in results]).sum(axis=0)
         self.dist_coordinate1 = np.array([item[2] for item in results]).sum(axis=0)
@@ -192,8 +337,11 @@ class Export:
             z = hdul[1]["Z"][:]
             eff_weights = hdul[1]["EFF_WEIGHTS"][:]
 
-            # TODO implement blinding support
-            distortion = hdul[2]["DISTORTION"][:]
+            blinding = header["BLINDING"] if "BLINDING" in header else "none"
+            distortion_name = "DISTORTION"
+            if "DISTORTION_BLIND" in hdul[2].get_colnames():
+                distortion_name += "_BLIND"
+            distortion = hdul[2][distortion_name][:]
             weights = hdul[2]["DISTORTION_WEIGHTS"][:]
 
         return (
@@ -205,6 +353,7 @@ class Export:
             eff_weights,
             num_pairs,
             num_pairs_used,
+            blinding,
         )
 
     def compute_covariance(self):
@@ -274,6 +423,10 @@ class Export:
         return covariance_smooth
 
     def write_correlation(self, global_config, settings):
+        xi = self._apply_blinding(self.mean_correlation.copy())
+        correlation_name = "DA_BLIND" if self._output_is_blinded() else "DA"
+        distortion_name = "DM_BLIND" if self._output_is_blinded() else "DM"
+
         output_file = self.output_directory / f"{self.name}-exp.fits.gz"
         results = fitsio.FITS(output_file, "rw", clobber=True)
 
@@ -301,7 +454,7 @@ class Export:
             },
             {
                 "name": "BLINDING",
-                "value": "placeholder",  # TODO Correct this once blinding implemented
+                "value": self.blinding,
                 "comment": "String specifying the blinding strategy",
             },
         ]
@@ -330,12 +483,12 @@ class Export:
                 self.coordinate1,
                 self.coordinate2,
                 self.z_grid,
-                self.mean_correlation,
+                xi,
                 self.covariance,
                 distortion,
                 self.num_pairs,
             ],
-            names=[*coordinate_names, "Z", "DA", "CO", "DM", "NB"],
+            names=[*coordinate_names, "Z", correlation_name, "CO", distortion_name, "NB"],
             comment=comment,
             header=header,
             extname="COR",
@@ -351,6 +504,7 @@ class Export:
         results.close()
 
     def write_distortion(self, global_config, settings):
+        distortion_name = "DM_BLIND" if self._output_is_blinded() else "DM"
         output_file = self.output_directory / f"dmat_{self.name}-exp.fits.gz"
         results = fitsio.FITS(output_file, "rw", clobber=True)
 
@@ -380,14 +534,14 @@ class Export:
             },
             {
                 "name": "BLINDING",
-                "value": "placeholder",  # TODO Correct this once blinding implemented
+                "value": self.blinding,
                 "comment": "String specifying the blinding strategy",
             },
         ]
 
         results.write(
             [self.dist_weights, self.distortion],
-            names=["WDM", "DM"],
+            names=["WDM", distortion_name],
             comment=["Sum of weights", "Distortion matrix"],
             header=header,
             extname="COR",