+++ lya_2pt/scripts/run_cf.py
@@ -3,6 +3,7 @@ import argparse
 import time
 from configparser import ConfigParser
 
+from lya_2pt.constants import ACCEPTED_BLIND_CORRELATION_TYPES
 from lya_2pt.interface import Interface
 
 
@@ -105,6 +106,12 @@ def main():
         required=False,
         help=("Number of cpus when running in parallel"),
     )
+    parser.add_argument(
+        "--blind-corr-type",
+        choices=ACCEPTED_BLIND_CORRELATION_TYPES,
+        default=None,
+        help="Correlation type used to select a DESI blinding template",
+    )
 
     args = parser.parse_args()
 
@@ -140,6 +147,8 @@ def main():
         "export-correlation": "True",
         "smooth-covariance": str(not args.no_smooth_cov),
     }
+    if args.blind_corr_type is not None:
+        config["export"]["blind-corr-type"] = args.blind_corr_type
 
     print("Initializing")
     total_t1 = time.time()